git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Prerelease/BLCore;C19344

+++ Source/ApplicationServices/2Gis.Erm.BLCore.Aggregates/Common/Generics/ICreateAggregateRepository.cs
@@ -7,6 +7,6 @@ namespace DoubleGis.Erm.BLCore.Aggregates.Common.Generics
     public interface ICreateAggregateRepository<in TEntity> : IUnknownAggregateSpecificOperation<CreateIdentity>
         where TEntity : class, IEntity, IEntityKey
     {
-        long Create(TEntity entity);
+        int Create(TEntity entity);
     }
 }
